Changing Recipe to use Jodit editor
