+++ src/App.js
@@ -14,7 +14,6 @@ import ManageUsers from './pages/ManageUsers';
 import EditUser from './pages/EditUser'
 import CreateUser from './pages/CreateUser'
 import EditRecipe from './pages/EditRecipe'
-import RTRecipeTest from './pages/RTRecipeTest';
 import { ToastContainer } from 'react-toastify';
 import "react-toastify/dist/ReactToastify.css";
 
@@ -42,7 +41,6 @@ function App() {
             <Route path="/edituser" element={<EditUser />} />
             <Route path="/createuser" element={<CreateUser />} />
             <Route path="/editrecipe" element={<EditRecipe />} />
-            <Route path="/rtrecipetest" element={<RTRecipeTest />} />
           </Routes>
         </div>
       </BrowserRouter>
